+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Performance:
 
 Other:
 - add `grunt-contrib-connect` to `Gruntfile.js` that allows to run an ad-hoc http://localhost:8080/ server with command `grunt connect`
+- bind `cellProperties` as `this` in `cells` callback
 
 ## [0.9.4](https://github.com/warpech/jquery-handsontable/tree/v0.9.4) (Jun 7, 2013)